refactor: firebase_optionをinfraディレクトリに移動 #18
